+++ src/components/Contact/ContactForm/ValidatedContactForm/ValidatedContactForm.js
@@ -10,7 +10,7 @@ const contactSchema = Yup.object().shape({
   email: Yup.string().email().required(),
   phone: Yup.string(),
   reason: Yup.string().required(),
-  message: Yup.string().min(50).max(1000).required()
+  message: Yup.string().min(100).max(2000).required()
 })
 
 export const ValidatedContactForm = ({ onSubmit }) => {